app/requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
requests==2.32.3
gunicorn==23.0.0